requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Faker==18.9.0
idna==3.4
interchange==2021.0.4
monotonic==1.6
neo4j==5.8.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pansi==2020.7.3
petl==1.7.12
psycopg2==2.9.6
py2neo==2021.2.3
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.2
